Reworked stack status
