+++ src/Exceptionless.Web/ClientApp/src/lib/components/data-table/index.ts
@@ -5,8 +5,6 @@ import FacetedFilter from './data-table-faceted-filter.svelte';
 import PageSize from './data-table-page-size.svelte';
 import Pagination from './data-table-pagination.svelte';
 
-import StatusCell from './data-table-status-cell.svelte';
-
 export {
 	Root,
 	Body,
@@ -14,13 +12,11 @@ export {
 	FacetedFilter,
 	PageSize,
 	Pagination,
-	StatusCell,
 	//
 	Root as DataTable,
 	Body as DataTableBody,
 	Toolbar as DataTableToolbar,
 	FacetedFilter as DataTableFacetedFilter,
 	PageSize as DataTablePageSize,
-	Pagination as DataTablePagination,
-	StatusCell as DataTableStatusCell
+	Pagination as DataTablePagination
 };
