+++ src/Exceptionless.Web/ClientApp/src/lib/components/events/table/EventsUserIdentitySummaryCell.svelte
