+++ src/Exceptionless.Web/ClientApp/src/lib/components/events/table/StackUsersSummaryCell.svelte
