+++ src/Exceptionless.Web/ClientApp/src/lib/components/ui/command/command-input.svelte
@@ -3,15 +3,13 @@
 	import { MagnifyingGlass } from 'radix-icons-svelte';
 	import { cn } from '$lib/utils';
 
-	type $$Props = CommandPrimitive.InputProps;
-
 	let className: string | undefined | null = undefined;
 	export { className as class };
 	export let value: string = '';
 </script>
 
-<div class="flex items-center border-b px-3" data-cmdk-input-wrapper="">
-	<MagnifyingGlass class="mr-2 h-4 w-4 shrink-0 opacity-50" />
+<div class="flex items-center px-3 border-b" data-cmdk-input-wrapper="">
+	<MagnifyingGlass class="w-4 h-4 mr-2 opacity-50 shrink-0" />
 	<CommandPrimitive.Input
 		class={cn(
 			'flex h-10 w-full rounded-md bg-transparent py-3 text-sm outline-none placeholder:text-muted-foreground disabled:cursor-not-allowed disabled:opacity-50',
